requirements.txt
torch>=2.0.0
transformers>=4.30.0
Pillow>=10.0.0
gradio>=4.0.0
numpy>=1.24.0
nltk>=3.8.0
sentencepiece>=0.1.99
tqdm>=4.65.0
